chore: Extract useTasks hook from new Project page
