+++ app/assets/js/models/tasks/index.tsx
@@ -6,6 +6,7 @@ import { Status } from "turboui/src/TaskBoard/types";
 import { parseContent, richContentToString } from "turboui/RichContent";
 
 export type { Task } from "@/api";
+export { useTasksForTurboUi } from "./useTasksForTurboUi";
 
 export {
   getTask,
